python3: add bluez5-native to DEPENDS conditionally
When bluetooth is in DISTRO_FEATURES, Python should enable support for Bluetooth sockets
in the socket library. To enable AF_BLUETOOTH socket family and its protocols, bluetooth.h
header needs to be available at build-time.

With this change a Python application can open and use Bluetooth sockets natively without
using 3rd party packages like pybluez.

An example is the following line:
    sock = socket.socket(socket.AF_BLUETOOTH, socket.SOCK_STREAM, socket.BTPROTO_RFCOMM)

With this patch the code will not fail due to missing symbols.

Signed-off-by: Guðni Már Gilbert <gudni.m.g@gmail.com>