bluez5: disable aics tests
Temporarily disable aics tests as they can fail
depending on how the tests are executed. Sometimes they pass,
sometimes they fail. The issue has been observed since BlueZ 5.72 to 5.80

Starting with BlueZ 5.80, the tests began failing when using the
ptest-runner script. This is not a new issue in BlueZ 5.80 which is
why the test is disabled with this commit until a solution is found.

Signed-off-by: Guðni Már Gilbert <gudni.m.g@gmail.com>